requirements.txt
boto3
awscli
pypdf
langchain-openai 
streamlit
faiss-cpu
langchain_community